Split firecrest scheduler implementation
